pyproject.toml
[tool.poetry]
name = "sbom_workshop"
version = "0.1.0"
description = ""
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cyclonedx-python-lib = ">=3.1,<4.0"
lief = {version = ">=0.12,<1.0", optional = true}
loguru = ">=0.6,<1.0"
packageurl-python = ">=0.10,<1.0"
pip-audit = ">=2.4,<3.0"
psutil = ">=5.9,<6.0"
pyparsing = ">=3.0,<4.0"
sh = ">=2.0,<3.0"
typer = ">=0.7,<1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1"
flake8 = "^6.0"
mypy = "^1.1"
pre-commit = "^3.2"
pycodestyle = "^2.10"
pytest = "^7.2"
types-setuptools = "^67.6"
vcrpy = "^4.2.1"

[tool.poetry.extras]
lief = ["lief"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sbom-workshop-cli = "sbom_workshop.cli:app"
